42 - Camera (3)
- 클릭 후, 목적지 Pos(vLookAt)을 넘어서서 떨리는 현상 해결
-> 시간 0.5f (초기화 값) 동안에 해당 fDistance (클릭 위치까지의 거리)를 카메라가 이동하도록 수정 완료
=> 시간 값을 주고, 해당 시간 (예시, 0.5f)를 넘어서면 m_vCurLookAt = m_vLookAt (목적지)로 강제 세팅함으로써 해결
(*) 시간 되면, 속도가 점점 느려져서 도착하도록 만들어보기

- 각종 warning (타입 변환 UINT) 수정

+++ WindowsGame/WindowsGame/CAnimation.cpp
@@ -94,7 +94,7 @@ void CAnimation::Create(const wstring& _strName, CTexture* _pTex
 	{
 		frm.fDuration = _fDuration;
 		frm.vSlice = _vSliceSize;
-		frm.vLT = _vLT + (_vStep * (int)i);
+		frm.vLT = _vLT + (_vStep * (float)i);
 
 		m_vecFrm.push_back(frm);
 	}
